DomainFilterPlugin: prevent exception if no config file when running the proxy
Even though the "config" parameter is mandatory I was able to somehow
start the proxy without ".config" being present in "self.flags"
(which is an argparse Namespace instance).

Not having a config file will trigger an exception later (when
the Configuration instance tries to load the file) but that can be
handled separately.